--- EssentialApp/EssentialApp/FeedLoaderPresentationAdapter.swift
@@ -1,35 +1,0 @@
-//
-//  Copyright © 2019 Essential Developer. All rights reserved.
-//
-
-import Combine
-import EssentialFeed
-import EssentialFeediOS
-
-final class FeedLoaderPresentationAdapter: FeedViewControllerDelegate {
-	private let feedLoader: () -> FeedLoader.Publisher
-	private var cancellable: Cancellable?
-	var presenter: LoadResourcePresenter<[FeedImage], FeedViewAdapter>?
-	
-	init(feedLoader: @escaping () -> FeedLoader.Publisher) {
-		self.feedLoader = feedLoader
-	}
-	
-	func didRequestFeedRefresh() {
-		presenter?.didStartLoading()
-		
-		cancellable = feedLoader()
-			.dispatchOnMainQueue()
-			.sink(
-				receiveCompletion: { [weak self] completion in
-					switch completion {
-					case .finished: break
-						
-					case let .failure(error):
-						self?.presenter?.didFinishLoading(with: error)
-					}
-				}, receiveValue: { [weak self] feed in
-					self?.presenter?.didFinishLoading(with: feed)
-				})
-	}
-}
