+++ EssentialApp/EssentialApp/FeedUIComposer.swift
@@ -14,7 +14,7 @@ public final class FeedUIComposer {
 		feedLoader: @escaping () -> FeedLoader.Publisher,
 		imageLoader: @escaping (URL) -> FeedImageDataLoader.Publisher
 	) -> FeedViewController {
-		let presentationAdapter = FeedLoaderPresentationAdapter(feedLoader: feedLoader)
+		let presentationAdapter = LoadResourcePresentationAdapter<[FeedImage], FeedViewAdapter>(loader: feedLoader)
 		
 		let feedController = makeFeedViewController(
 			delegate: presentationAdapter,
